broke down minishell.c into auxilary files + norminette check half-OK
