+++ Makefile
@@ -23,6 +23,9 @@ VALGRIND		=	valgrind --leak-check=full \
 					--track-fds=yes
 
 FILES			=	minishell.c \
+					exit_routines.c \
+					hashtable.c \
+					init_routines.c \
 					utils_test.c
 
 HEADERS			=	$(PATH_INCLUDES)minishell.h \
